scuba: Move ScubaConfig.process_command() to ScubaContext classmethod
