+++ scuba/config.py
@@ -272,8 +272,6 @@ class ScubaAlias:
 
         return cls(name, script, image, entrypoint, environment, shell, as_root, docker_args)
 
-class ScubaContext:
-    pass
 
 class ScubaConfig:
     def __init__(self, **data):
@@ -347,85 +345,6 @@ class ScubaConfig:
     @property
     def docker_args(self):
         return self._docker_args
-
-
-    def process_command(self, command, image=None, shell=None):
-        '''Processes a user command using aliases
-
-        Arguments:
-            command     A user command list (e.g. argv)
-            image       Override the image from .scuba.yml
-            shell       Override the shell from .scuba.yml
-
-        Returns: A ScubaContext object with the following attributes:
-            script: a list of command line strings
-            image: the docker image name to use
-        '''
-        result = ScubaContext()
-        result.script = None
-        result.image = None
-        result.entrypoint = self.entrypoint
-        result.environment = self.environment.copy()
-        result.shell = self.shell
-        result.as_root = False
-        result.docker_args = self.docker_args
-
-        if command:
-            alias = self.aliases.get(command[0])
-            if not alias:
-                # Command is not an alias; use it as-is.
-                result.script = [shell_quote_cmd(command)]
-            else:
-                # Using an alias
-                # Does this alias override the image and/or entrypoint?
-                if alias.image:
-                    result.image = alias.image
-                if alias.entrypoint is not None:
-                    result.entrypoint = alias.entrypoint
-                if alias.shell is not None:
-                    result.shell = alias.shell
-                if alias.as_root:
-                    result.as_root = True
-
-                if isinstance(alias.docker_args, OverrideMixin) or result.docker_args is None:
-                    result.docker_args = alias.docker_args
-                elif alias.docker_args is not None:
-                    result.docker_args.extend(alias.docker_args)
-
-                # Merge/override the environment
-                if alias.environment:
-                    result.environment.update(alias.environment)
-
-                if len(alias.script) > 1:
-                    # Alias is a multiline script; no additional
-                    # arguments are allowed in the scuba invocation.
-                    if len(command) > 1:
-                        raise ConfigError('Additional arguments not allowed with multi-line aliases')
-                    result.script = alias.script
-
-                else:
-                    # Alias is a single-line script; perform substituion
-                    # and add user arguments.
-                    command.pop(0)
-                    result.script = [alias.script[0] + ' ' + shell_quote_cmd(command)]
-
-            result.script = flatten_list(result.script)
-
-        # If a shell was given on the CLI, it should override the shell set by
-        # the alias or top-level config
-        if shell:
-            result.shell = shell
-
-        # If an image was given, it overrides what might have been set by an alias
-        if image:
-            result.image = image
-
-        # If the image was still not set, then try to get it from the confg,
-        # which will raise a ConfigError if it is not set
-        if not result.image:
-            result.image = self.image
-
-        return result
 
 
 def load_config(path):
